DM-11603 fix highlight and select issue for 'scatter' type chart (fireflyScatter and scatter)
