+++ src/firefly/js/charts/dataTypes/FireflyScatter.js
@@ -104,9 +104,11 @@ function fetchData(chartId, traceNum, tablesource) {
             const yColumn = getColumn(tableModel, get(mappings, 'y'));
             const yUnit = get(yColumn, 'units', '');
 
+            const traceData = get(getChartData(chartId), `data.${traceNum}`);
+
             addOtherChanges({changes, xLabel, xTipLabel, xUnit, yLabel, yTipLabel, yUnit, chartId, traceNum});
             dispatchChartUpdate({chartId, changes});
-            dispatchChartHighlighted({chartId, highlighted: getPointIdx(highlightedRow)});   // update highlighted point in chart
+            dispatchChartHighlighted({chartId, highlighted: getPointIdx(traceData, highlightedRow)});   // update highlighted point in chart
             updateSelected(chartId, selectInfo);
         }
     }).catch(
